Move isTruncated paginator trait to own decorator
