+++ CHANGELOG.next.toml
@@ -34,9 +34,3 @@ message = "Fix S3 ListParts API paginator infinite loop."
 references = ["aws-sdk-rust#1143"]
 meta = { "breaking" = false, "tada" = false, "bug" = true }
 author = "landonxjames"
-
-[[smithy-rs]]
-message = "Fix S3 ListParts API paginator infinite loop."
-references = ["aws-sdk-rust#1143"]
-meta = { "breaking" = false, "tada" = false, "bug" = true, "target" = "client"}
-author = "landonxjames"
