+++ aws/sdk-codegen/src/main/kotlin/software/amazon/smithy/rustsdk/customize/s3/S3Decorator.kt
@@ -28,7 +28,6 @@ import software.amazon.smithy.rust.codegen.client.smithy.generators.OperationGen
 import software.amazon.smithy.rust.codegen.client.smithy.generators.OperationSection
 import software.amazon.smithy.rust.codegen.client.smithy.protocols.ClientRestXmlFactory
 import software.amazon.smithy.rust.codegen.client.smithy.traits.IncompatibleWithStalledStreamProtectionTrait
-import software.amazon.smithy.rust.codegen.client.smithy.traits.IsTruncatedPaginatorTrait
 import software.amazon.smithy.rust.codegen.core.rustlang.Writable
 import software.amazon.smithy.rust.codegen.core.rustlang.rustBlockTemplate
 import software.amazon.smithy.rust.codegen.core.rustlang.rustTemplate
@@ -65,7 +64,6 @@ class S3Decorator : ClientCodegenDecorator {
         setOf(
             ShapeId.from("com.amazonaws.s3#CopyObject"),
         )
-    private val operationsWithIsTruncatedPaginator = setOf(ShapeId.from("com.amazonaws.s3#ListPartsOutput"))
 
     override fun protocols(
         serviceId: ShapeId,
@@ -91,9 +89,6 @@ class S3Decorator : ClientCodegenDecorator {
             }.letIf(operationsIncompatibleWithStalledStreamProtection.contains(shape.id)) {
                 logger.info("Adding IncompatibleWithStalledStreamProtection trait to $it")
                 (it as OperationShape).toBuilder().addTrait(IncompatibleWithStalledStreamProtectionTrait()).build()
-            }.letIf(isInIsTruncatedList(shape)) {
-                logger.info("Adding IsTruncatedPaginator trait to $it")
-                (it as StructureShape).toBuilder().addTrait(IsTruncatedPaginatorTrait()).build()
             }
         }
             // the model has the bucket in the path
@@ -170,10 +165,6 @@ class S3Decorator : ClientCodegenDecorator {
 
     private fun isInInvalidXmlRootAllowList(shape: Shape): Boolean {
         return shape.isStructureShape && invalidXmlRootAllowList.contains(shape.id)
-    }
-
-    private fun isInIsTruncatedList(shape: Shape): Boolean {
-        return shape.isStructureShape && operationsWithIsTruncatedPaginator.contains(shape.id)
     }
 }
